requirements.txt
# Status LLMs Research Dependencies

# Core API libraries
openai>=1.0.0
anthropic>=0.8.0
google-genai>=1.0.0
requests>=2.28.0

# Utilities
python-dotenv>=1.0.0

# Optional: Alternative unified API
# aimlapi>=1.0.0  # Uncomment if using AI/ML API